Refactor and rename new field
